Allow for customization of the schedule start time
If Cron is not active when a schedule occurs, it's possible we'll miss
a task that we need to run. We've seen this happen when schedules
coincide with a deploy.

This commit adds a configuration option which will allow us to set the
initial start time. We can "catch up" missed tasks by passing the
shutdown time on creation.