Do not use the epoch as the initial "now" time
We've previously added support for setting the "epoch" for our cron;
that is, when we should start scheduling from. However, we were also
using the epoch as the initial time for computing how long to wait
until the next job should run. This resulted in drift of `now - epoch`.

With this commit the initial time is _always_ the current time; the
epoch defaults to the current time, unless overridden.